Create Java Program to Check Whether the Character is Vowel or Consonant
For any given character, we need to check if it is a vowel or a consonant. As we know, vowels are ‘a’, ‘e’, ‘i’, ‘o’, ‘u’ and all the other characters (i.e. ‘b’, ‘c’, ‘d’, ‘f’ …..) are consonants.

Lightbox

Examples:  

Input : char = 'r'
Output : Consonant

Input : char = 'e'
Output : Vowel
Here, in the below implementation we will check if the stated character corresponds to any of the five vowels. And if it matches, “Vowel” is printed, else “Consonant” is printed.