Create Java Program for factorial of a number
Factorial of a non-negative integer, is multiplication of all integers smaller than or equal to n. In this article, we will learn how to write a program for the factorial of a number in Java.

Formulae for Factorial
n! = n * (n-1) * (n-2) * (n-3) * ........ * 1
Factorial of a Number in Java
Example of Factorial in Java
6! == 6*5*4*3*2*1 = 720. 
5! == 5*4*3*2*1 = 120
4! == 4*3*2*1 = 24